+++ test/_get/get.core.specifics.spec.js
@@ -200,62 +200,4 @@ describe('Specific Cases', function() {
         });
     });
 
-    it('should use path syntax to retrieve values out, with double ranges.', function(done) {
-        var spy = sinon.spy();
-        var model = new Model({
-            source: new LocalDataSource(
-                Cache(), {
-                    onGet: spy
-                })
-        });
-
-        model.
-            get('genreList[0][0..1, 3..4].summary').
-            doAction(function(x) {
-                var call = spy.getCall(0);
-                var arg = call.args[1].map(function(path) {
-                    return path.map(function(x) {
-                        if (Array.isArray(x)) {
-                            return x.concat();
-                        }
-                        return x;
-                    });
-                });
-                testRunner.compare([['genreList', 0, [0, 1, 3, 4], 'summary']], arg);
-
-                testRunner.compare({
-                    json: {
-                        genreList: {
-                            0: {
-                                0: {
-                                    summary: {
-                                        title: 'House of Cards',
-                                        url: '/movies/1234'
-                                    }
-                                },
-                                1: {
-                                    summary: {
-                                        'title': 'Terminator 3',
-                                        'url': '/movies/766'
-                                    }
-                                },
-                                3: {
-                                    summary: {
-                                        'title': 'Commando',
-                                        'url': '/movies/6420'
-                                    }
-                                },
-                                4: {
-                                    summary: {
-                                        'title': 'Additional Title 0',
-                                        'url': '/movies/0'
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }, x);
-            }).
-            subscribe(noOp, done, done);
-    });
 });
