+++ test/falcor/get/get.cache-only.spec.js
@@ -1,4 +1,4 @@
-var falcor = require("./../../../lib/");
+var falcor = require("./../../../lib");
 var Model = falcor.Model;
 var Cache = require('../../data/Cache');
 var Expected = require('../../data/expected');
@@ -30,65 +30,18 @@ describe('Cache Only', function() {
                     }
                 });
 
-            model.get(["user", "name"]).toPathValues().subscribe(function(pathValue) {
-                value = pathValue;
-            });
+            model.
+                get(["user", "name"]).
+                subscribe(function(out) {
+                    value = pathValue;
+                });
 
             if (value === undefined) {
                 throw new Error("Value not retrieved from cache, despite the fact that it has not expired");
             }
         });
     });
-
-    describe('Allow Arrays in JSON Graph branch position', function() {
-        it('get should go right through arrays in branch position', function() {
-            var model = new falcor.Model({
-                cache: {
-                    users: [
-                        {
-                            name: "Jim",
-                            age: 23
-                        },
-                        {
-                            name: "John",
-                            age: 44
-                        },
-                        {
-                            name: "Derek",
-                            age: 22
-                        }
-                    ]
-                }
-            });
-
-            model.get(["users", {length:3}, ["name","age"]]).toPathValues().toArray().subscribe(function(vals) {
-                // console.log(JSON.stringify(vals));
-                testRunner.compare(
-                    JSON.stringify(vals),
-                    '[{"path":["users",0,"name"],"value":"Jim"},{"path":["users",0,"age"],"value":23},{"path":["users",1,"name"],"value":"John"},{"path":["users",1,"age"],"value":44},{"path":["users",2,"name"],"value":"Derek"},{"path":["users",2,"age"],"value":22}]',
-                    'JSON Graphs That contain arrays in the bridge deposition return the same thing as those that contain maps in the branch position when given the same query');
-            });
-        });
-    });
-
-    describe('toJSON', function() {
-        it('should get a value from falcor.', function(done) {
-            var model = new Model({cache: Cache()});
-            var expected = Expected.Values().direct.AsPathMap.values[0];
-            var next = false;
-            model.
-                get(['videos', 1234, 'summary']).
-                toJSON().
-                doAction(function(x) {
-                    testRunner.compare(expected, x);
-                    next = true;
-                }, noOp, function() {
-                    testRunner.compare(true, next, 'Expect to be onNext at least 1 time.');
-                }).
-                subscribe(noOp, done, done);
-        });
-    });
-    describe('_toJSONG', function() {
+    describe('PathMap', function() {
         it('should get a value from falcor.', function(done) {
             var model = new Model({cache: Cache()});
             var expected = Expected.Values().direct.AsJSONG.values[0];
@@ -104,73 +57,21 @@ describe('Cache Only', function() {
                 }).
                 subscribe(noOp, done, done);
         });
-        it('should get JSONGraph value-types out that were user created.', function(done) {
-            var model = new Model({
-                cache: {
-                    foo: 5,
-                    bar: {
-                        $type: 'atom',
-                        $hello: 'world',
-                        value: 5
-                    }
-                }
-            });
-            var onNext = sinon.spy();
-            model.
-                get('foo', 'bar').
-                _toJSONG().
-                doAction(onNext, noOp, function() {
-                    expect(onNext.calledOnce).to.be.ok;
-                    expect(onNext.getCall(0).args[0]).to.deep.equals({
-                        jsonGraph: {
-                            foo: 5,
-                            bar: {
-                                $type: 'atom',
-                                $hello: 'world',
-                                $size: 51,
-                                value: 5
-                            }
-                        },
-                        paths: [['foo'], ['bar']]
-                    });
-                }).
-                subscribe(noOp, done, done);
-        });
-        it('should get nulls out as JSONG without boxed values.', function(done) {
-            var model = new Model({
-                cache: {
-                    foo: null
-                }
-            });
-            var onNext = sinon.spy();
-            model.
-                get('foo').
-                _toJSONG().
-                doAction(onNext, noOp, function() {
-                    expect(onNext.calledOnce).to.be.ok;
-                    expect(onNext.getCall(0).args[0]).to.deep.equals({
-                        jsonGraph: {
-                            foo: null
-                        },
-                        paths: [['foo']]
-                    });
-                }).
-                subscribe(noOp, done, done);
-        });
     });
-    describe('toPathValues', function() {
+
+    describe('_toJSONG', function() {
         it('should get a value from falcor.', function(done) {
             var model = new Model({cache: Cache()});
-            var expected = Expected.Values().direct.AsValues.values[0];
-            var next = 0;
+            var expected = Expected.Values().direct.AsJSONG.values[0];
+            var next = false;
             model.
                 get(['videos', 1234, 'summary']).
-                toPathValues().
+                _toJSONG().
                 doAction(function(x) {
                     testRunner.compare(expected, x);
-                    ++next;
+                    next = true;
                 }, noOp, function() {
-                    testRunner.compare(1, next, 'Expect to be onNext 1 time.');
+                    testRunner.compare(true, next, 'Expect to be onNext at least 1 time.');
                 }).
                 subscribe(noOp, done, done);
         });
